+++ db/records/operations/__init__.py
@@ -1,0 +1,5 @@
+from db.records.operations.delete import *  # noqa
+from db.records.operations.group import *  # noqa
+from db.records.operations.insert import *  # noqa
+from db.records.operations.select import *  # noqa
+from db.records.operations.update import *  # noqa
