+++ db/records/utils.py
@@ -1,0 +1,5 @@
+def create_col_objects(table, column_list):
+    return [
+        table.columns[col] if type(col) == str else table.columns[col.name]
+        for col in column_list
+    ]
