+++ db/tables/utils.py
@@ -1,4 +1,5 @@
 from sqlalchemy import Table, MetaData
+from sqlalchemy.inspection import inspect
 
 from db import columns
 
@@ -27,3 +28,10 @@ def get_enriched_column_table(raw_sa_table, engine=None):
 
 def get_empty_table(name):
     return Table(name, MetaData())
+
+
+def get_primary_key_column(table):
+    primary_key_list = list(inspect(table).primary_key)
+    # We do not support getting by composite primary keys
+    assert len(primary_key_list) == 1
+    return primary_key_list[0]
