+++ db/tests/records/test_grouping.py
@@ -5,14 +5,14 @@ import pytest
 from sqlalchemy import select
 from sqlalchemy_filters import apply_sort, apply_filters
 
-from db import records
+from db.records.operations.group import get_group_counts
 from db.records.exceptions import BadGroupFormat, GroupFieldNotFound
 
 
 def test_get_group_counts_str_field(filter_sort_table_obj):
     filter_sort, engine = filter_sort_table_obj
     group_by = ["varchar"]
-    counts = records.get_group_counts(filter_sort, engine, group_by)
+    counts = get_group_counts(filter_sort, engine, group_by)
     assert len(counts) == 101
     assert ("string1",) in counts
 
@@ -20,7 +20,7 @@ def test_get_group_counts_str_field(filter_sort_table_obj):
 def test_get_group_counts_col_field(filter_sort_table_obj):
     filter_sort, engine = filter_sort_table_obj
     group_by = [filter_sort.c.varchar]
-    counts = records.get_group_counts(filter_sort, engine, group_by)
+    counts = get_group_counts(filter_sort, engine, group_by)
     assert len(counts) == 101
     assert ("string1",) in counts
 
@@ -28,7 +28,7 @@ def test_get_group_counts_col_field(filter_sort_table_obj):
 def test_get_group_counts_mixed_str_col_field(filter_sort_table_obj):
     filter_sort, engine = filter_sort_table_obj
     group_by = ["varchar", filter_sort.c.numeric]
-    counts = records.get_group_counts(filter_sort, engine, group_by)
+    counts = get_group_counts(filter_sort, engine, group_by)
     assert len(counts) == 101
     assert ("string1", 1) in counts
 
@@ -45,8 +45,7 @@ def test_get_group_counts_limit_offset_ordering(roster_table_obj, limit, offset)
     roster, engine = roster_table_obj
     order_by = [{"field": "Grade", "direction": "desc", "nullslast": True}]
     group_by = [roster.c["Grade"]]
-    counts = records.get_group_counts(roster, engine, group_by, limit=limit,
-                                      offset=offset, order_by=order_by)
+    counts = get_group_counts(roster, engine, group_by, limit=limit, offset=offset, order_by=order_by)
     query = select(group_by[0])
     query = apply_sort(query, order_by)
     with engine.begin() as conn:
@@ -85,7 +84,7 @@ count_values_test_list = itertools.chain(*[
 @pytest.mark.parametrize("group_by", count_values_test_list)
 def test_get_group_counts_count_values(roster_table_obj, group_by):
     roster, engine = roster_table_obj
-    counts = records.get_group_counts(roster, engine, group_by)
+    counts = get_group_counts(roster, engine, group_by)
 
     cols = [roster.c[f] for f in group_by]
     with engine.begin() as conn:
@@ -112,7 +111,7 @@ filter_values_test_list = itertools.chain(*[
 def test_get_group_counts_filter_values(roster_table_obj, filter_by):
     roster, engine = roster_table_obj
     group_by = ["Student Name"]
-    counts = records.get_group_counts(roster, engine, group_by, filters=filter_by)
+    counts = get_group_counts(roster, engine, group_by, filters=filter_by)
 
     cols = [roster.c[f] for f in group_by]
     query = select(*cols)
@@ -138,4 +137,4 @@ exceptions_test_list = [
 def test_get_group_counts_exceptions(filter_sort_table_obj, group_by, exception):
     filter_sort, engine = filter_sort_table_obj
     with pytest.raises(exception):
-        records.get_group_counts(filter_sort, engine, group_by)
+        get_group_counts(filter_sort, engine, group_by)
