+++ mathesar/tests/api/test_record_api.py
@@ -2,7 +2,7 @@ import json
 from unittest.mock import patch
 
 import pytest
-from db import records
+from db.records import operations as record_operations
 from db.records.exceptions import BadGroupFormat, GroupFieldNotFound
 from sqlalchemy_filters.exceptions import BadFilterFormat, BadSortFormat, FilterFieldNotFound, SortFieldNotFound
 
@@ -70,7 +70,7 @@ def test_record_list_filter(create_table, client):
     json_filter_list = json.dumps(filter_list)
 
     with patch.object(
-        records, "get_records", side_effect=records.get_records
+        record_operations, "get_records", side_effect=record_operations.get_records
     ) as mock_get:
         response = client.get(
             f'/api/v0/tables/{table.id}/records/?filters={json_filter_list}'
@@ -93,7 +93,7 @@ def test_record_list_filter_duplicates(create_table, client):
     ]
     json_filter_list = json.dumps(filter_list)
 
-    with patch.object(records, "get_records") as mock_get:
+    with patch.object(record_operations, "get_records") as mock_get:
         client.get(f'/api/v0/tables/{table.id}/records/?filters={json_filter_list}')
     assert mock_get.call_args is not None
     assert mock_get.call_args[1]['filters'] == filter_list
@@ -121,7 +121,7 @@ def _test_filter_with_added_columns(table, client, columns_to_add, operators_and
             json_filter_list = json.dumps(filter_list)
 
             with patch.object(
-                records, "get_records", side_effect=records.get_records
+                record_operations, "get_records", side_effect=record_operations.get_records
             ) as mock_get:
                 response = client.get(
                     f'/api/v0/tables/{table.id}/records/?filters={json_filter_list}'
@@ -172,7 +172,7 @@ def test_record_list_sort(create_table, client):
     json_order_by = json.dumps(order_by)
 
     with patch.object(
-        records, "get_records", side_effect=records.get_records
+        record_operations, "get_records", side_effect=record_operations.get_records
     ) as mock_get:
         response = client.get(
             f'/api/v0/tables/{table.id}/records/?order_by={json_order_by}'
@@ -197,7 +197,7 @@ def _test_record_list_group(table, client, group_count_by, expected_groups):
     query_str = f'group_count_by={json_group_count_by}&order_by={json_order_by}'
 
     with patch.object(
-        records, "get_group_counts", side_effect=records.get_group_counts
+        record_operations, "get_group_counts", side_effect=record_operations.get_group_counts
     ) as mock_get:
         response = client.get(f'/api/v0/tables/{table.id}/records/?{query_str}')
         response_data = response.json()
@@ -397,7 +397,7 @@ def test_record_list_filter_exceptions(create_table, client, exception):
     table_name = f"NASA Record List {exception.__name__}"
     table = create_table(table_name)
     filter_list = json.dumps([{"field": "Center", "op": "is_null"}])
-    with patch.object(records, "get_records", side_effect=exception):
+    with patch.object(record_operations, "get_records", side_effect=exception):
         response = client.get(
             f'/api/v0/tables/{table.id}/records/?filters={filter_list}'
         )
@@ -412,7 +412,7 @@ def test_record_list_sort_exceptions(create_table, client, exception):
     table_name = f"NASA Record List {exception.__name__}"
     table = create_table(table_name)
     order_by = json.dumps([{"field": "Center", "direction": "desc"}])
-    with patch.object(records, "get_records", side_effect=exception):
+    with patch.object(record_operations, "get_records", side_effect=exception):
         response = client.get(
             f'/api/v0/tables/{table.id}/records/?order_by={order_by}'
         )
@@ -427,7 +427,7 @@ def test_record_list_group_exceptions(create_table, client, exception):
     table_name = f"NASA Record List {exception.__name__}"
     table = create_table(table_name)
     group_by = json.dumps(["Center"])
-    with patch.object(records, "get_group_counts", side_effect=exception):
+    with patch.object(record_operations, "get_group_counts", side_effect=exception):
         response = client.get(
             f'/api/v0/tables/{table.id}/records/?group_count_by={group_by}'
         )
